movie-details/requirements.txt
python-dotenv>=1.0.0
Requests>=2.31.0
asyncio>=3.4.3
